[ZIPFLDR] Extraction is now in its own thread (#2261)
Extraction happens now in its own thread, allowing for cancelling in the middle of extraction or moving the window during the extraction.

File extraction was previously happening in the main thread, freezing the whole window until the end. Now when the user clicks "Next", a new thread is created and the extraction starts from there, allowing to move the window during the extraction, or to cancel the extraction in the middle. (Note: after clicking "Cancel", extraction continues until the current file is extracted).

CORE-14934