[SHELL32] Don't hardcode My Documents path for My Music/Pictures/Videos folders in some translations (#2354)
Don't hardcode My Documents path in some translations of My Music, My Pictures and My Videos folders in Shell32. It fixes incorrect location of that folders when install ReactOS as Workstation.
Currently they are located in one more My Documents subfolder, which is, in its turn, in the correct My Documents directory. After this fix, they are directly in the 2nd one, as they should be.

CORE-16698