[DRWTSN32] Recognize EXCEPTION_WINE_STUB (#2514)
DrWtsn32.exe recognizes several standard exception codes, and prints their names in the crash log. However, EXCEPTION_WINE_STUB is not in this list, and as such it is treated as an unknown error and printed as "--". Having this error code given a name at the start of the crash dump is useful, as it allows me to identify calls to unimplemented issues more readily.