[BROWSEUI] Display an error message if the path or drive don't exist (CORE-10989)
The return result of ParseNow was never checked, but contained useful information on what type of error may have occurred.  If the error is an issue with the path not existing the explorer will now show an error message notifying the user of the problem instead of silently failing.
Note:  Shell32 does not properly return ERROR_INVALID_DRIVE when the drive path does not exist.  For that reason, it is still possible to browse to a non-existent drive.  However, once shell32 is patched this code will work properly, as it checks for both ERROR_INVALID_DRIVE and ERROR_FILE_NOT_FOUND.