[IMM32] ValidateHwndNoErr(): Apply ReactOS-specific handle table array fixup. (#3937) CORE-17741
This function (which should belong to the user32 helpers) depends on the
internal format of the USER handles table, that is ReactOS-specific. Its
pointer value is stored in the win32k/user32 shared info section, and
thus cannot be converted to user-space and stored back into that section.
Therefore it needs to be retrieved and manually converted when the function
is called.